+++ app/files/page.tsx
@@ -22,9 +22,9 @@ import { Progress } from "@/components/ui/progress";
 import { Badge } from "@/components/ui/badge";
 import ThemeToggle from "@/components/theme-toggle";
 import Link from "next/link";
-import { FileDropzone } from "@/components/file-dropzone";
+import { FileDropzone } from "@/components/files/file-dropzone";
 import { CaptchaModal } from "@/components/captcha-modal";
-import { LinkResultModal } from "@/components/link-result-modal";
+import { LinkResultModal } from "@/components/rooms/link-result-modal";
 import { AnimatePresence, motion } from "motion/react";
 import { Slider } from "@/components/ui/slider";
 import Script from "next/script";
