+++ components/rooms/link-result-modal.tsx
